+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## TBD
 
+- Added mastery for Femtobyte
 - Reduced lag spikes even further by tweaking mod dll export settings
 - Femtobyte now flips tables on placement
 - Made detonation sound for Scotsman a constant volume instead of scaling with number of projectiles